Added EventPageBottomDrawer
