+++ src/screens/EventPage/index.tsx
@@ -1,5 +1,6 @@
 export { default as EventPageDetail } from './EventPageDetail';
 export { default as EventPageCounter } from './EventPageCounter';
+export { default as EventPageBottomDrawer } from './EventPageBottomDrawer';
 
 import EventPage from './EventPage';
